requirements.txt
# Flask and web framework
Flask==3.0.0
Werkzeug==3.0.0

# AI and LangChain
anthropic==0.71.0
langchain==1.0.0
langchain-anthropic==1.0.0
langchain-community==0.4
langchain-core==1.0.0

# Utilities
python-dotenv==1.0.0
requests==2.32.5

# TOTP Authentication (for app.py)
pyotp==2.9.0
qrcode==8.2
Pillow==10.0.0

# Production server
gunicorn==21.2.0
